requirements.txt
sounddevice
scipy
openai-whisper
pyttsx3
speechrecognition
requests
beautifulsoup4
